Major correction on repositories

+++ src/main/java/com/tacospasa/msdata/Repository/AccessRepository.java
@@ -3,5 +3,5 @@ package com.tacospasa.msdata.Repository;
 import com.tacospasa.msdata.Entity.AccessEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccessRepository extends JpaRepository<AccessEntity, Integer> {
+public interface AccessRepository extends JpaRepository<AccessEntity, Long> {
 }
