+++ src/main/java/com/tacospasa/msdata/Repository/MenuCategoryRepository.java
@@ -3,5 +3,5 @@ package com.tacospasa.msdata.Repository;
 import com.tacospasa.msdata.Entity.MenuCategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuCategoryRepository extends JpaRepository<MenuCategoryEntity, Integer> {
+public interface MenuCategoryRepository extends JpaRepository<MenuCategoryEntity, Long> {
 }
