+++ src/main/java/com/tacospasa/msdata/Repository/PresentationRepository.java
@@ -3,5 +3,5 @@ package com.tacospasa.msdata.Repository;
 import com.tacospasa.msdata.Entity.PresentationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PresentationRepository extends JpaRepository<PresentationEntity,Integer> {
+public interface PresentationRepository extends JpaRepository<PresentationEntity,Long> {
 }
