+++ src/main/java/com/tacospasa/msdata/Repository/RoleRepository.java
@@ -2,6 +2,8 @@ package com.tacospasa.msdata.Repository;
 
 import com.tacospasa.msdata.Entity.RoleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
-public interface RoleRepository extends JpaRepository<RoleEntity, Integer> {
+public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
 }
