+++ src/main/java/com/tacospasa/msdata/Service/MenuCategoryService.java
@@ -6,10 +6,10 @@ import com.tacospasa.msdata.Entity.SubCategoryEntity;
 import java.util.List;
 
 public interface MenuCategoryService {
-    MenuCategoryEntity getMenuCategoryById(int id);
+    MenuCategoryEntity getMenuCategoryById(Long id);
     List<MenuCategoryEntity> getAllMenuCategories();
     MenuCategoryEntity createNewMenuCategory(MenuCategoryEntity menuCategory);
-    MenuCategoryEntity addSubCategories(List<SubCategoryEntity> subCategories, int id);
-    MenuCategoryEntity addSubCategory(SubCategoryEntity subCategory, int id);
+    MenuCategoryEntity addSubCategories(List<SubCategoryEntity> subCategories, Long id);
+    MenuCategoryEntity addSubCategory(SubCategoryEntity subCategory, Long id);
 
 }
